just like this joke, this rock is a little flat

+++ twc/embedded/renderd/RenderScript.py
@@ -1157,7 +1157,7 @@ class AudioClip(AudioRenderable):
         return
 
     def duration(self):
-        return int(self.file.get_length()*30)
+        return (self.file.get_length()*30)
         return
 
     def size(self):
